Rebased

+++ ubelt/util_io.py
@@ -1,27 +1,47 @@
 # -*- coding: utf-8 -*-
+"""
+Functions for reading and writing files on disk.
+
+:func:`writeto` and :func:`readfrom` wrap ``open().write()`` and
+``open().read()`` and primarily serve to indicate that the type of data being
+written and read is unicode text.
+
+:func:`delete` wraps :func:`os.unlink` and :func:`shutil.rmtree` and does not
+throw an error if the file or directory does not exist. It also contains
+workarounds for win32 issues with :mod:`shutil`.
+"""
 from __future__ import absolute_import, division, print_function, unicode_literals
 import six
+import sys
 import os
-from os.path import exists, join  # NOQA
+from os.path import exists
 
 
-def writeto(fpath, to_write, aslines=False, mode='w', verbose=None):
+__all__ = [
+    'readfrom', 'writeto', 'touch', 'delete',
+]
+
+
+def writeto(fpath, to_write, aslines=False, verbose=None):
     r"""
-    Writes text to a file. Automatically encodes text as utf8.
+    Writes (utf8) text to a file.
 
     Args:
-        fpath (str): file path
+        fpath (str | PathLike): file path
         to_write (str): text to write (must be unicode text)
         aslines (bool): if True to_write is assumed to be a list of lines
         verbose (bool): verbosity flag
-        mode (unicode): (default = u'w')
-        n (int):  (default = 2)
 
-    CommandLine:
-        python -m ubelt.util_io writeto --verbose
+    Note:
+        You probably should use `open(<fpath>).write(<to_write>)` instead.
+        This function exists as a convenience for writing in Python2. After
+        2020-01-01, we may consider deprecating the function.
+
+        NOTE: It turns out that `open(<fpath>).write(<to_write>)` does not work
+        in pypy. See `https://pypy.org/compat.html`. This is a strong argument
+        for keeping this function.
 
     Example:
-        >>> from ubelt.util_io import *  # NOQA
         >>> import ubelt as ub
         >>> dpath = ub.ensure_app_cache_dir('ubelt')
         >>> fpath = dpath + '/' + 'testwrite.txt'
@@ -35,7 +55,6 @@ def writeto(fpath, to_write, aslines=False, mode='w', verbose=None):
         >>> assert read_ == to_write
 
     Example:
-        >>> from ubelt.util_io import *  # NOQA
         >>> import ubelt as ub
         >>> dpath = ub.ensure_app_cache_dir('ubelt')
         >>> fpath = dpath + '/' + 'testwrite2.txt'
@@ -51,59 +70,86 @@ def writeto(fpath, to_write, aslines=False, mode='w', verbose=None):
     if verbose:
         print('Writing to text file: %r ' % (fpath,))
 
-    with open(fpath, mode) as file_:
+    with open(fpath, 'wb') as file:
         if aslines:
-            file_.writelines(to_write)
+            to_write = map(_ensure_bytes , to_write)
+            file.writelines(to_write)
         else:
-            # Ensure python2 writes in bytes
-            if six.PY2 and isinstance(to_write, unicode):  # NOQA
-                to_write = to_write.encode('utf8')  # nocover
-            file_.write(to_write)
+            # convert to bytes for writing
+            bytes = _ensure_bytes(to_write)
+            file.write(bytes)
+
+
+if six.PY2:  # nocover
+    def _ensure_bytes(text):
+        if isinstance(text, unicode):  # NOQA
+            text = text.encode('utf8')
+        return text
+else:
+    def _ensure_bytes(text):
+        """ ensures text is in a suitable format for writing """
+        return text.encode('utf8')
 
 
 def readfrom(fpath, aslines=False, errors='replace', verbose=None):
-    r"""
-    Reads text from a file. Automatically returns utf8.
+    """
+    Reads (utf8) text from a file.
+
+    Note:
+        You probably should use `open(<fpath>).read()` instead.
+        This function exists as a convenience for writing in Python2. After
+        2020-01-01, we may consider deprecating the function.
 
     Args:
-        fpath (str): file path
+        fpath (str | PathLike): file path
         aslines (bool): if True returns list of lines
         verbose (bool): verbosity flag
 
     Returns:
         str: text from fpath (this is unicode)
-
-    Ignore:
-        x = b'''/whaleshark_003_fors\xc3\xb8g.wmv" />\r\n'''
-        ub.writeto('foo.txt', x)
-        y = ub.readfrom('foo.txt')
-        y.encode('utf8') == x
     """
     if verbose:
-        print('[util_io] * Reading text file: %r ' % (fpath,))
+        print('Reading text file: %r ' % (fpath,))
     if not exists(fpath):
         raise IOError('File %r does not exist' % (fpath,))
-    with open(fpath, 'rb') as file_:
+    with open(fpath, 'rb') as file:
         if aslines:
             text = [line.decode('utf8', errors=errors)
-                    for line in file_.readlines()]
+                    for line in file.readlines()]
+            if sys.platform.startswith('win32'):  # nocover
+                # fix line endings on windows
+                text = [
+                    line[:-2] + '\n' if line.endswith('\r\n') else line
+                    for line in text
+                ]
         else:
-            text = file_.read().decode('utf8', errors=errors)
+            text = file.read().decode('utf8', errors=errors)
     return text
 
 
-def touch(fname, mode=0o666, dir_fd=None, **kwargs):
-    r"""
+def touch(fpath, mode=0o666, dir_fd=None, verbose=0, **kwargs):
+    """
     change file timestamps
 
     Works like the touch unix utility
+
+    Args:
+        fpath (str | PathLike): name of the file
+        mode (int): file permissions (python3 and unix only)
+        dir_fd (file): optional directory file descriptor. If specified, fpath
+            is interpreted as relative to this descriptor (python 3 only).
+        verbose (int): verbosity
+        **kwargs : extra args passed to ``os.utime`` (python 3 only).
+
+    Returns:
+        str: path to the file
 
     References:
         https://stackoverflow.com/questions/1158076/implement-touch-using-python
 
     Example:
-        >>> from ubelt.util_io import *  # NOQA
         >>> import ubelt as ub
+        >>> from os.path import join
         >>> dpath = ub.ensure_app_cache_dir('ubelt')
         >>> fpath = join(dpath, 'touch_file')
         >>> assert not exists(fpath)
@@ -111,27 +157,36 @@ def touch(fname, mode=0o666, dir_fd=None, **kwargs):
         >>> assert exists(fpath)
         >>> os.unlink(fpath)
     """
+    if verbose:
+        print('Touching file {}'.format(fpath))
     if six.PY2:  # nocover
-        with open(fname, 'a'):
-            os.utime(fname, None)
+        with open(fpath, 'a'):
+            os.utime(fpath, None)
     else:
         flags = os.O_CREAT | os.O_APPEND
-        with os.fdopen(os.open(fname, flags=flags, mode=mode, dir_fd=dir_fd)) as f:
-            os.utime(f.fileno() if os.utime in os.supports_fd else fname,
+        with os.fdopen(os.open(fpath, flags=flags, mode=mode, dir_fd=dir_fd)) as f:
+            os.utime(f.fileno() if os.utime in os.supports_fd else fpath,
                      dir_fd=None if os.supports_fd else dir_fd, **kwargs)
+    return fpath
 
 
 def delete(path, verbose=False):
     """
     Removes a file or recursively removes a directory.
-    If a path does not exist, then this is a noop
+    If a path does not exist, then this is does nothing.
 
     Args:
-        path (str): file or directory to remove
+        path (str | PathLike): file or directory to remove
+        verbose (bool): if True prints what is being done
 
-    Doctest:
+    SeeAlso:
+        :mod:`send2trash` - A cross-platform Python package for sending files
+            to the trash instead of irreversibly deleting them.
+            https://github.com/hsoft/send2trash
+
+    Example:
         >>> import ubelt as ub
-        >>> from os.path import join, exists
+        >>> from os.path import join
         >>> base = ub.ensure_app_cache_dir('ubelt', 'delete_test')
         >>> dpath1 = ub.ensuredir(join(base, 'dir'))
         >>> ub.ensuredir(join(base, 'dir', 'subdir'))
@@ -146,26 +201,53 @@ def delete(path, verbose=False):
         >>> assert not exists(fpath1)
         >>> ub.delete(dpath1)
         >>> assert not any(map(exists, (dpath1, fpath1, fpath2)))
+
+    Example:
+        >>> import ubelt as ub
+        >>> from os.path import exists, join
+        >>> dpath = ub.ensure_app_cache_dir('ubelt', 'delete_test2')
+        >>> dpath1 = ub.ensuredir(join(dpath, 'dir'))
+        >>> fpath1 = ub.touch(join(dpath1, 'to_remove.txt'))
+        >>> assert exists(fpath1)
+        >>> ub.delete(dpath)
+        >>> assert not exists(fpath1)
     """
-    import shutil
     if not os.path.exists(path):
-        if verbose:  # nocover
-            print('Not deleting non-existant path="{}"'.format(path))
+        # if the file does exists and is not a broken link
+        if os.path.islink(path):
+            if verbose:  # nocover
+                print('Deleting broken link="{}"'.format(path))
+            os.unlink(path)
+        elif os.path.isdir(path):  # nocover
+            # Only on windows will a file be a directory and not exist
+            if verbose:
+                print('Deleting broken directory link="{}"'.format(path))
+            os.rmdir(path)
+        elif os.path.isfile(path):  # nocover
+            # This is a windows only case
+            if verbose:
+                print('Deleting broken file link="{}"'.format(path))
+            os.unlink(path)
+        else:
+            if verbose:  # nocover
+                print('Not deleting non-existant path="{}"'.format(path))
     else:
-        if os.path.isfile(path) or os.path.islink(path):
+        if os.path.islink(path):
+            if verbose:  # nocover
+                print('Deleting symbolic link="{}"'.format(path))
+            os.unlink(path)
+        elif os.path.isfile(path):
             if verbose:  # nocover
                 print('Deleting file="{}"'.format(path))
             os.unlink(path)
         elif os.path.isdir(path):
             if verbose:  # nocover
                 print('Deleting directory="{}"'.format(path))
-            shutil.rmtree(path)
-
-
-if __name__ == '__main__':
-    r"""
-    CommandLine:
-        python -m ubelt.util_io
-    """
-    import ubelt as ub  # NOQA
-    ub.doctest_package()
+            if sys.platform.startswith('win32'):  # nocover
+                # Workaround bug that prevents shutil from working if
+                # the directory contains junctions
+                from ubelt import _win32_links
+                _win32_links._win32_rmtree(path, verbose=verbose)
+            else:
+                import shutil
+                shutil.rmtree(path)
